move get_test_dbconf to be a shared test helper. use in both metadata and ui_backend service tests
